move indices_where to searching, copy_array to creation
