+++ ivy/functional/backends/tensorflow/searching.py
@@ -5,6 +5,10 @@ import ivy
 import tensorflow as tf
 
 
+# Array API Standard #
+# ------------------ #
+
+
 def argmax(
     x: Union[tf.Tensor, tf.Variable],
     /,
@@ -48,3 +52,21 @@ def where(
 ) -> Union[tf.Tensor, tf.Variable]:
     x1, x2 = ivy.promote_types_of_inputs(x1, x2)
     return tf.experimental.numpy.where(condition, x1, x2)
+
+
+# Extra #
+# ----- #
+
+
+def indices_where(
+    x: Union[tf.Tensor, tf.Variable],
+    *,
+    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
+) -> Union[tf.Tensor, tf.Variable]:
+    where_x = tf.experimental.numpy.where(x)
+    if len(where_x) == 1:
+        return tf.expand_dims(where_x[0], -1)
+    res = tf.experimental.numpy.concatenate(
+        [tf.expand_dims(item, -1) for item in where_x], -1
+    )
+    return res
